TRIPLEX-820 Update Storybook configuration: replace MediaWidth story with OrderedList story and remove OrderedList story file to streamline component documentation.
