--- stories/OrderedList.stories.tsx
@@ -1,83 +1,0 @@
-import React from "react";
-import { Meta, StoryObj } from "@storybook/react";
-import { OrderedList } from "../src/components/OrderedList";
-import { Title, Description, Primary, Controls, Stories } from "@storybook/addon-docs/blocks";
-
-export default {
-    title: "Components/OrderedList",
-    component: OrderedList,
-    tags: ["autodocs"],
-    parameters: {
-        testRunner: { skip: true },
-        layout: "padded",
-        docs: {
-            description: {
-                component: `Ordered list for displaying sequential items.
-
-**Usage example:**
-
-\`\`\`tsx
-import { OrderedList } from "@sberbusiness/triplex-next/components/OrderedList";
-
-<OrderedList>
-    <OrderedList.Item>List item text;</OrderedList.Item>
-    <OrderedList.Item>List item text;</OrderedList.Item>
-    <OrderedList.Item>List item text.</OrderedList.Item>
-</OrderedList>
-\`\`\`
-                `,
-            },
-            page: () => (
-                <>
-                    <Title />
-                    <Description />
-                    <Controls of={Default} />
-                    <Primary />
-                    <Stories />
-                </>
-            ),
-        },
-    },
-    decorators: [
-        (Story) => (
-            <div style={{ maxWidth: "200px" }}>
-                <Story />
-            </div>
-        ),
-    ],
-} as Meta<typeof OrderedList>;
-
-type Story = StoryObj<typeof OrderedList>;
-
-export const Default: Story = {
-    name: "Default",
-    parameters: {
-        controls: { disable: true },
-    },
-    render: () => (
-        <OrderedList>
-            <OrderedList.Item>List item text;</OrderedList.Item>
-            <OrderedList.Item>List item text;</OrderedList.Item>
-            <OrderedList.Item>List item text.</OrderedList.Item>
-        </OrderedList>
-    ),
-};
-
-export const WithLongContent: Story = {
-    name: "With Long Content",
-    render: () => (
-        <OrderedList>
-            <OrderedList.Item>List item text;</OrderedList.Item>
-            <OrderedList.Item>List item text;</OrderedList.Item>
-            <OrderedList.Item>Extremely long list item that fits in two lines.</OrderedList.Item>
-        </OrderedList>
-    ),
-    parameters: {
-        controls: { disable: true },
-        docs: {
-            description: {
-                story: "List with items containing multiline text.",
-            },
-        },
-    },
-};
