+++ stories/OrderedList/examples/DefaultExample.tsx
@@ -1,0 +1,12 @@
+import React from "react";
+import { OrderedList } from "@sberbusiness/triplex-next";
+
+export const DefaultExample = () => (
+    <div style={{ maxWidth: "200px" }}>
+        <OrderedList>
+            <OrderedList.Item>List item text;</OrderedList.Item>
+            <OrderedList.Item>List item text;</OrderedList.Item>
+            <OrderedList.Item>List item text.</OrderedList.Item>
+        </OrderedList>
+    </div>
+);
